requirements.txt
aiohttp>=3.8.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
colorama>=0.4.6
boto3>=1.26.0
dnspython>=2.3.0
cryptography>=41.0.0
pyjwt>=2.8.0
urllib3>=2.0.0
tqdm>=4.65.0
psutil>=5.8.0
pyyaml>=6.0
flask>=3.0.0
fastapi>=0.100.0
uvicorn>=0.20.0
jinja2>=3.1.0
click>=8.0.0
asyncio-throttle>=1.0.0